requirements.txt
discord.py
pterodactyl.py